Rename BaseAttenuator to ReadOnlyAttenuator (#931)
